test: consolidate red-attacker tests into proper test files
- red-bundle-hash-cache-isolation: deleted (covered by namespace-isolation)
- red-namespace-algorithm-parity: 5 unique tests merged into namespace-isolation
- red-paths-namespace-paths: 8 unique tests merged into namespace-isolation
- red-hud-canonicalization: renamed to hud-canonicalization, duplicates removed
- red-legacy-job-backendnamespace: 6 tests merged into progress-store
